requirements.txt
apipkg==1.5; python_version != '3.3.*'
argh==0.26.2
astroid==2.0.4
atomicwrites==1.2.1; python_version != '3.3.*'
attrs==18.2.0
autopep8==1.4
bandit==1.5.1
certifi==2018.8.24
chardet==3.0.4
click==6.7
colorama==0.3.9; sys_platform == 'win32'
coverage==4.5.1
dataclasses==0.6
docopt==0.6.2
dparse==0.4.1
elib-config==2018.9.2.1
elib-run==2018.9.9.1
execnet==1.5.0; python_version != '3.3.*'
faker==0.9.0
flake8==3.5.0
future==0.16.0
gitchangelog==3.0.3
gitdb2==2.0.4
gitpython==2.1.11
hypothesis==3.71.1
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mockito==1.1.1
more-itertools==4.3.0
multidict==4.4.0; python_version >= '3.4.1'
mypy==0.620
packaging==17.1
pathtools==0.1.2
pbr==4.2.0
pefile==2018.8.8
pkginfo==1.4.2
pluggy==0.7.1; python_version != '3.3.*'
py==1.6.0; python_version != '3.3.*'
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==2.1.1
pyparsing==2.2.0
pytest-cache==1.0
pytest-cov==2.6.0
pytest-deadfixtures==2.0.1
pytest-faker==2.0.0
pytest-pycharm==0.5.0; os_name == 'nt'
pytest-repeat==0.7.0
pytest-vcr==0.3.0
pytest-watch==4.2.0
pytest==3.8.0
python-dateutil==2.7.3
pyyaml==3.13
requests-toolbelt==0.8.0; python_version != '3.3.*'
requests==2.19.1; python_version != '3.3.*'
safety==1.8.4
sarge==0.1.5.post0
setuptools-scm==3.1.0
six==1.11.0
smmap2==2.0.4
stevedore==1.29.0
text-unidecode==1.2
toml==0.9.6
tqdm==4.25.0; python_version != '3.0.*'
twine==1.11.0
typed-ast==1.1.0; python_version < '3.7' and implementation_name == 'cpython'
urllib3==1.23; python_version != '3.3.*'
vcrpy==1.13.0; python_version >= '3.4.1'
watchdog==0.9.0
wheel==0.31.1
wrapt==1.10.11
yarl==1.2.6; python_version >= '3.4'